// src/app/dashboard/interfaces/profile.ts
// Generated by https://quicktype.io

export interface Profile {
  data: Data;
}

export interface Data {
  id: string;
  first_name: string;
  last_name: string;
  email: string;
  phone: string;
  street: string;
  ext_num: string;
  int_num: string;
  suburb: string;
  city: string;
  state: string;
  zip_code: string;
  photo: string;
  created_at: string;
  updated_at: string;
}